Fixing bugs in EvtBTo{3pi,Kpipi} models.
In couple of places there were small bugs in generation of kinematics for
events as code used wrong daughter masses. Effect of the bug is not
completely clear due to the fact that code uses 5-vectors with mass squared
of the particle as 5th component (used typically instead of calculating
mass from 4-vector). This code will be removed soon, but want to have
version with known bugs fixed in single commit.