Prototype of ordering daughters in DDALITZ.
Modify way how we decide which decay we have and in which order daughters
were passed for DDALITZ decay model. In new version, we first sort daughter
ids and then compress all comparisons to single statement per mother and
daughters. Once code is correct for one ordering of daughters, it is
correct for all other orderings. In this commit flags 1 and 2 are
rewritten.